refactor: gameManagement->Game, Game->Pitch
