+++ src/main/java/baseball/Application.java
@@ -7,10 +7,11 @@ public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
 
-        GameManagement gm = new GameManagement();
-        while (!gm.isGameFinished()) {
-            gm.playGame(scanner);
-            gm.askReGame(scanner);
+        boolean onGame = true;
+        while (onGame) {
+            Game game = new Game();
+            game.play(scanner);
+            onGame = game.askReGame(scanner);
         }
     }
 }
