+++ src/main/java/baseball/Game.java
@@ -2,76 +2,68 @@ package baseball;
 
 import utils.ThreeDigitsUtils;
 
+import java.util.Scanner;
+
 public class Game {
-    private final String NOTHING = "낫싱";
-    private final String BALL = "볼";
-    private final String STRIKE = "스트라이크";
-    private int targetNumber;
+    private final String QUESTION_FOR_BASEBALL_NUMBER = "숫자를 입력해주세요 : ";
+    private final String QUESTION_FOR_RE_GAME = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
+    private final String CONGRATULATION_MESSAGE = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
+    private Pitch pitch;
 
-    public Game(int targetNumber) {
-        this.targetNumber = targetNumber;
+    public Game() {
+        int targetNumber = ThreeDigitsUtils.nextBaseballNumber();
+        pitch = new Pitch(targetNumber);
+        System.out.println(targetNumber);
     }
 
-    private int countStrike(int inputNumber) {
-        int[] targetArray = ThreeDigitsUtils.toIntArrayOfDigits(targetNumber);
-        int[] inputArray = ThreeDigitsUtils.toIntArrayOfDigits(inputNumber);
-
-        int count = 0;
-        for (int i = 0; i < 3; i++) {
-            if (targetArray[i] == inputArray[i]) {
-                count += 1;
+    public void play(Scanner scanner) {
+        while (true){
+            int guess = askNumber(scanner);
+            printHint(guess);
+            if (pitch.isCorrect(guess)) {
+                break;
             }
         }
-        return count;
     }
 
-    private int countOverlap(int inputNumber) {
-        int[] targetArray = ThreeDigitsUtils.toIntArrayOfDigits(targetNumber);
-        int[] inputArray = ThreeDigitsUtils.toIntArrayOfDigits(inputNumber);
 
-        int count = 0;
-        for (int i = 0; i < 3; i++) {
-            if (containInt(targetArray[i], inputArray)) {
-                count += 1;
-            }
+    private int askNumber(Scanner scanner) {
+        System.out.print(QUESTION_FOR_BASEBALL_NUMBER);
+        int inputInt = toInt(scanner.nextLine());
+        if (!ThreeDigitsUtils.isBaseballNumber(inputInt)) {
+            throw new IllegalArgumentException("Input should be Distinct Three Digits Integer");
         }
-        return count;
+        return inputInt;
     }
 
-    private boolean containInt(int number, int[] Array) {
-        for ( int a : Array) {
-            if ( a == number) {
-                return true;
-            }
+    private void printHint(int guess) {
+        String hint = pitch.getHint(guess);
+        System.out.println(hint);
+    }
+
+    public boolean askReGame(Scanner scanner) {
+        System.out.println(CONGRATULATION_MESSAGE);
+        System.out.println(QUESTION_FOR_RE_GAME);
+        return answerReGame(scanner.nextLine());
+    }
+
+    private boolean answerReGame(String inputString) {
+        int inputInt = toInt(inputString);
+        if (inputInt == 1) {
+            return true;
         }
-        return false;
-    }
-
-    private String stringifyHint(int strikeCount, int ballCount) {
-        if (strikeCount == 0 && ballCount == 0){
-            return NOTHING;
+        if (inputInt == 2) {
+            return false;
         }
-        String message = "";
-        if (ballCount > 0) {
-            message = Integer.toString(ballCount) + BALL;
+        throw new IllegalArgumentException("Input should be 1 or 2");
+    }
+
+    private int toInt(String inputString) {
+        String string = inputString.strip();
+        try {
+            return Integer.parseInt(string);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("Input should be only Integer");
         }
-        if (strikeCount > 0) {
-            message = message + Integer.toString(strikeCount) + STRIKE;
-        }
-        return message;
-    }
-
-    public String getHint(int inputNumber) {
-        int strikeCount = countStrike(inputNumber);
-        int ballCount = countOverlap(inputNumber) - strikeCount;
-        return stringifyHint(strikeCount, ballCount);
-    }
-
-    public int getTargetNumber() {
-        return targetNumber;
-    }
-
-    public boolean isCorrect(int guess) {
-        return guess == targetNumber;
     }
 }
